Finish

+++ app/Http/Controllers/GamesController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Game;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class GamesController extends Controller
 {
@@ -14,6 +15,7 @@ class GamesController extends Controller
     public function index()
     {
         $games = Game::all();
+        $games = Game::paginate(5);
         return view('games.index', ['games' => $games]);
     }
 
@@ -30,13 +32,15 @@ class GamesController extends Controller
      */
     public function store(Request $request)
     {
+        $data = $request->all();
         $newGame = new Game();
-        $newGame->name = $request->name;
-        $newGame->genre = $request->genre;
-        $newGame->duration = $request->duration;
-        $newGame->price = $request->price;
+        $newGame->name = $data["name"];
+        $newGame->genre = $data["genre"];
+        $newGame->duration = $data["duration"];
+        $newGame->price = $data["price"];
+        $newGame->user_id = Auth::user()->id;
         $newGame->save();
-        return redirect()->route('games.index');
+        return redirect()->route('games.index')->with('success_create', $newGame);
     }
 
     /**
@@ -51,8 +55,11 @@ class GamesController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Game $game)
+    public function edit(Game $game, Request $request)
     {
+        if (auth()->user()->id !== $game->user_id) {
+            abort(403, 'Unauthorized action.');
+        }
         return view('games.edit', ['game' => $game]);
     }
 
@@ -61,20 +68,29 @@ class GamesController extends Controller
      */
     public function update(Request $request, Game $game)
     {
-        $game->name = $request->name;
-        $game->genre = $request->genre;
-        $game->duration = $request->duration;
-        $game->price = $request->price;
+
+        if (auth()->user()->id !== $game->user_id) {
+            abort(403, 'Unauthorized action.');
+        }
+        $data = $request->all();
+        $game->name = $data["name"];
+        $game->genre = $data["genre"];
+        $game->duration = $data["duration"];
+        $game->price = $data["price"];
+        $game->user_id = Auth::user()->id;
         $game->save();
-        return redirect()->route('games.index');
+        return redirect()->route('games.index')->with('success_update', $game);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Game $game)
+    public function destroy(Game $game, Request $request)
     {
+        if (auth()->user()->id != $game->user_id) {
+            abort(403, 'Unauthorized action.');
+        }
         $game->delete();
-        return redirect()->route('games.index');
+        return redirect()->route('games.index')->with('success_deleted', $game);
     }
 }
